IPToSAT Enabled if JSON available
This enables the player if the user has synchronized channels.
I find that OpenSPA disables it by default and there are two consistent reasons for it not to:
If there is a list of synchronized channels, there is no excuse to disable it since they force the user first to change the boolean and second to restart the interface.
If user install the channel list from the plugin (files channels lists + files iptosat ( the plugin restart IU) and after, you will have everything you need to start viewing in your TV the subscription. the user does not have to do anything else.